# bin/main/br/com/generated/api/AlertsApi.kt
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.10.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
*/
package br.com.generated.api

import br.com.generated.model.AlertResponse
import br.com.generated.model.ErrorResponse
import br.com.generated.model.UpdateAlertStatusRequest
import org.springframework.http.HttpStatus
import org.springframework.http.MediaType
import org.springframework.http.ResponseEntity

import org.springframework.web.bind.annotation.*
import org.springframework.validation.annotation.Validated
import org.springframework.web.context.request.NativeWebRequest
import org.springframework.beans.factory.annotation.Autowired

import jakarta.validation.constraints.DecimalMax
import jakarta.validation.constraints.DecimalMin
import jakarta.validation.constraints.Email
import jakarta.validation.constraints.Max
import jakarta.validation.constraints.Min
import jakarta.validation.constraints.NotNull
import jakarta.validation.constraints.Pattern
import jakarta.validation.constraints.Size
import jakarta.validation.Valid

import kotlin.collections.List
import kotlin.collections.Map

@RestController
@Validated
interface AlertsApi {


    @RequestMapping(
            method = [RequestMethod.GET],
            value = ["/v1/alerts/{alertId}"],
            produces = ["application/json"]
    )
    fun getAlertById( @PathVariable("alertId") alertId: java.util.UUID): ResponseEntity<AlertResponse>


    @RequestMapping(
            method = [RequestMethod.GET],
            value = ["/v1/alerts"],
            produces = ["application/json"]
    )
    fun searchAlerts( @Valid @RequestParam(value = "transactionId", required = false) transactionId: kotlin.String?, @Valid @RequestParam(value = "ruleId", required = false) ruleId: java.util.UUID?, @Valid @RequestParam(value = "page", required = false, defaultValue = "0") page: kotlin.Int, @Max(100)  @Valid @RequestParam(value = "size", required = false, defaultValue = "20") size: kotlin.Int): ResponseEntity<List<AlertResponse>>


    @RequestMapping(
            method = [RequestMethod.PATCH],
            value = ["/v1/alerts/{alertId}/status"],
            produces = ["application/json"],
            consumes = ["application/json"]
    )
    fun updateAlertStatus( @PathVariable("alertId") alertId: java.util.UUID, @Valid @RequestBody updateAlertStatusRequest: UpdateAlertStatusRequest): ResponseEntity<AlertResponse>
}