# bin/main/br/com/generated/api/ContextualScreeningApi.kt
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.10.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
*/
package br.com.generated.api

import br.com.generated.model.AnalystDecisionRequest
import br.com.generated.model.AnalystDecisionResponse
import br.com.generated.model.ContextualScreeningRequest
import br.com.generated.model.ContextualScreeningResponse
import br.com.generated.model.ErrorResponse
import org.springframework.http.HttpStatus
import org.springframework.http.MediaType
import org.springframework.http.ResponseEntity

import org.springframework.web.bind.annotation.*
import org.springframework.validation.annotation.Validated
import org.springframework.web.context.request.NativeWebRequest
import org.springframework.beans.factory.annotation.Autowired

import jakarta.validation.constraints.DecimalMax
import jakarta.validation.constraints.DecimalMin
import jakarta.validation.constraints.Email
import jakarta.validation.constraints.Max
import jakarta.validation.constraints.Min
import jakarta.validation.constraints.NotNull
import jakarta.validation.constraints.Pattern
import jakarta.validation.constraints.Size
import jakarta.validation.Valid

import kotlin.collections.List
import kotlin.collections.Map

@RestController
@Validated
interface ContextualScreeningApi {


    @RequestMapping(
            method = [RequestMethod.POST],
            value = ["/v1/rules/contextual-screening/evaluate"],
            produces = ["application/json"],
            consumes = ["application/json"]
    )
    fun evaluateContextualScreening( @Valid @RequestBody contextualScreeningRequest: ContextualScreeningRequest): ResponseEntity<ContextualScreeningResponse>


    @RequestMapping(
            method = [RequestMethod.POST],
            value = ["/v1/rules/contextual-screening/decisions"],
            produces = ["application/json"],
            consumes = ["application/json"]
    )
    fun registerAnalystDecision( @Valid @RequestBody analystDecisionRequest: AnalystDecisionRequest): ResponseEntity<AnalystDecisionResponse>
}