# bin/main/br/com/generated/api/DecisionExecutionApi.kt
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.10.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
*/
package br.com.generated.api

import br.com.generated.model.Decision
import br.com.generated.model.DecisionExecutionResponse
import br.com.generated.model.PagedDecisionExecutionResponse
import org.springframework.http.HttpStatus
import org.springframework.http.MediaType
import org.springframework.http.ResponseEntity

import org.springframework.web.bind.annotation.*
import org.springframework.validation.annotation.Validated
import org.springframework.web.context.request.NativeWebRequest
import org.springframework.beans.factory.annotation.Autowired

import jakarta.validation.constraints.DecimalMax
import jakarta.validation.constraints.DecimalMin
import jakarta.validation.constraints.Email
import jakarta.validation.constraints.Max
import jakarta.validation.constraints.Min
import jakarta.validation.constraints.NotNull
import jakarta.validation.constraints.Pattern
import jakarta.validation.constraints.Size
import jakarta.validation.Valid

import kotlin.collections.List
import kotlin.collections.Map

@RestController
@Validated
interface DecisionExecutionApi {


    @RequestMapping(
            method = [RequestMethod.GET],
            value = ["/v1/decision/executions/{id}"],
            produces = ["application/json"]
    )
    fun getExecutionById( @PathVariable("id") id: java.util.UUID): ResponseEntity<DecisionExecutionResponse>


    @RequestMapping(
            method = [RequestMethod.GET],
            value = ["/v1/decision/executions"],
            produces = ["application/json"]
    )
    fun searchExecutions( @Valid @RequestParam(value = "transactionId", required = false) transactionId: kotlin.String?, @Valid @RequestParam(value = "ruleId", required = false) ruleId: java.util.UUID?, @Valid @RequestParam(value = "decision", required = false) decision: Decision?, @Valid @RequestParam(value = "traceId", required = false) traceId: kotlin.String?,@Min(0) @Valid @RequestParam(value = "page", required = false, defaultValue = "0") page: kotlin.Int,@Min(1) @Max(100)  @Valid @RequestParam(value = "size", required = false, defaultValue = "20") size: kotlin.Int): ResponseEntity<PagedDecisionExecutionResponse>
}