# bin/main/static/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Decision Engine — Rule Platform API
  description: |
    # Decision Engine — Plataforma de Regras de PLD

    API da plataforma de decisão para Prevenção à Lavagem de Dinheiro (PLD).
    Desacopla a lógica de **detecção** (Screening Context) da lógica de **decisão**
    (quando gerar alertas, com quais critérios, baseado em quais fatos).

    ## Arquitetura

    ```
    ┌─────────────────┐     DetectionEvent      ┌──────────────────┐    DecisionMadeEvent    ┌───────────────┐
    │  Screening      │ ──────────────────────▶ │  Decision        │ ─────────────────────▶ │  Alert        │
    │  Context (MF09) │                         │  Context         │                        │  Context      │
    │                 │                         │                  │                        │               │
    │  Detecta        │                         │  Avalia regras   │                        │  Cria alertas │
    │  keywords       │                         │  Decide ações    │                        │  com status   │
    └─────────────────┘                         └──────────────────┘                        └───────────────┘
    ```

    ## Fluxo de Decisão

    ```
    1. Transação PIX recebida
    2. KeywordScreeningService detecta termo restrito → publica DetectionEvent
    3. Decision Engine consome evento:
       a) Verifica idempotência (transactionId + ruleId)
       b) Carrega configuração ativa da regra
...
    5. Publica DecisionMadeEvent
...

...

...

...

...
servers:
...
paths:
...
  /v1/decision/executions:
    get:
      tags: [Decision Execution]
      summary: Consultar execuções de decisão
      description: |
        Consulta o registro de execuções do Decision Engine para auditoria.
        **Pelo menos um filtro é obrigatório** (transactionId, ruleId, decision, ou traceId).

        Cada execução é imutável — nunca alterada após criação.
        Inclui: decisão tomada, fatos avaliados, expressões satisfeitas/falhas,
        tempo de execução e traceId para correlação com logs.

        Paginação: page size padrão 20, máximo 100.
      operationId: searchExecutions
      parameters:
        - name: transactionId
          in: query
          required: false
          description: Filtrar por ID da transação PIX
          schema:
            type: string
          example: "TX-001"
        - name: ruleId
          in: query
          required: false
          description: Filtrar por ID da regra (UUID)
          schema:
            type: string
            format: uuid
        - name: decision
          in: query
          required: false
          description: Filtrar por tipo de decisão
          schema:
            $ref: '#/components/schemas/Decision'
          example: ALERT
        - name: traceId
          in: query
          required: false
          description: Buscar por traceId (correlação com logs). Retorna resultado único.
          schema:
            type: string
        - name: page
          in: query
          required: false
          description: Número da página (0-indexed)
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: size
          in: query
          required: false
          description: Tamanho da página (mín 1, máx 100, padrão 20)
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Resultado paginado de execuções (ou único para traceId)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedDecisionExecutionResponse'
              examples:
                alertExecution:
                  summary: Execução com decisão ALERT
                  value:
                    content:
                      - id: "880e8400-e29b-41d4-a716-446655440000"
                        transactionId: "TX-001"
                        ruleId: "550e8400-e29b-41d4-a716-446655440000"
                        decision: "ALERT"
                        actions: ["GENERATE_ALERT"]
                        facts:
                          keywordMatched: true
                          customerRisk: "AR"
                        matchedExpressions:
                          - factName: "keywordMatched"
                            operator: "EQUALS"
                            expectedValue: true
                            actualValue: true
                            satisfied: true
                            justification: "Fact 'keywordMatched' (true) é igual a true"
                        failedExpressions: []
                        configurationVersion: 3
                        executionTimeMs: 12
                        traceId: "trace-uuid-123"
                        timestamp: "2025-01-15T10:35:00Z"
                    page: 0
                    size: 20
                    totalElements: 1
                    totalPages: 1
        '400':
          description: Nenhum filtro informado (pelo menos um é obrigatório)
  /v1/decision/executions/{id}:
    get:
      tags: [Decision Execution]
      summary: Buscar execução por ID
      operationId: getExecutionById
      parameters:
        - name: id
          in: path
          required: true
          description: UUID da execução
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Execução encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecisionExecutionResponse'
        '404':
          description: Execução não encontrada

  # ═══════════════════════════════════════════════════════════════════════════════
  # ALERTS — Gerenciamento de alertas
  # ═══════════════════════════════════════════════════════════════════════════════
  /v1/alerts:
    get:
      tags: [Alerts]
      summary: Consultar alertas
      description: |
        Consulta alertas gerados pelo Decision Engine.
        **Pelo menos um filtro é obrigatório** (transactionId ou ruleId).

        Alertas são criados automaticamente quando:
        - Decision Engine decide ALERT
        - Action configurada é GENERATE_ALERT
        - DecisionMadeEvent é publicado com sucesso

        Idempotência: apenas 1 alerta por (transactionId, ruleId).
      operationId: searchAlerts
      parameters:
        - name: transactionId
          in: query
          required: false
          description: Filtrar por ID da transação
          schema:
            type: string
          example: "TX-001"
        - name: ruleId
          in: query
          required: false
          description: Filtrar por ID da regra (paginado)
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Lista de alertas ou resultado paginado
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AlertResponse'
              examples:
                openAlert:
                  summary: Alerta aberto
                  value:
                    - id: "990e8400-e29b-41d4-a716-446655440000"
                      transactionId: "TX-001"
                      ruleId: "550e8400-e29b-41d4-a716-446655440000"
                      customerId: "CUST-42"
                      facts:
                        keywordMatched: true
                        customerRisk: "AR"
                      configurationVersion: 3
                      traceId: "trace-uuid-123"
                      actions: ["GENERATE_ALERT"]
                      explanation:
                        traceId: "trace-uuid-123"
                        decision: "ALERT"
                      status: "OPEN"
                      createdAt: "2025-01-15T10:35:00Z"
                      updatedAt: "2025-01-15T10:35:00Z"
        '400':
          description: Nenhum filtro informado
  /v1/alerts/{alertId}:
    get:
      tags: [Alerts]
      summary: Buscar alerta por ID
      operationId: getAlertById
      parameters:
        - name: alertId
          in: path
          required: true
          description: UUID do alerta
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Alerta encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertResponse'
        '404':
          description: Alerta não encontrado
  /v1/alerts/{alertId}/status:
    patch:
      tags: [Alerts]
      summary: Atualizar status do alerta
      description: |
        Atualiza o status de um alerta seguindo a state machine:

        ```
        ┌──────┐          ┌──────────────┐          ┌────────┐
        │ OPEN │ ────────▶│ UNDER_REVIEW │ ────────▶│ CLOSED │
        │      │          │              │          │        │
        └──────┘          └──────┬───────┘          └────────┘
                                 │
                                 │          ┌────────────────┐
                                 └────────▶│ FALSE_POSITIVE │
                                            │                │
                                            └────────────────┘
        ```

        **Transições válidas:**
        - OPEN → UNDER_REVIEW
        - UNDER_REVIEW → CLOSED
        - UNDER_REVIEW → FALSE_POSITIVE

        **Estados terminais** (sem transição possível):
        - CLOSED
        - FALSE_POSITIVE
      operationId: updateAlertStatus
      parameters:
        - name: alertId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAlertStatusRequest'
            examples:
              startReview:
                summary: Iniciar revisão
                value:
                  status: "UNDER_REVIEW"
              closeAlert:
                summary: Fechar alerta (confirmado)
                value:
                  status: "CLOSED"
              falsePositive:
                summary: Marcar como falso positivo
                value:
                  status: "FALSE_POSITIVE"
      responses:
        '200':
          description: Status atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertResponse'
        '404':
          description: Alerta não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: |
            Transição de status inválida (ex.: OPEN → CLOSED, CLOSED → qualquer)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidTransition:
                  value:
                    timestamp: "2025-01-15T10:40:00Z"
                    status: 422
                    error: "Unprocessable Entity"
                    message: "Transição inválida: OPEN → CLOSED"
                    details: null

  # ═══════════════════════════════════════════════════════════════════════════════
  # KEYWORD SCREENING — Detecção de termos restritos (MF09)
  # ═══════════════════════════════════════════════════════════════════════════════
...
  /v1/rules/contextual-screening/evaluate:
    post:
      tags: [Contextual Screening]
      summary: Avaliar contexto da transação via LLM
      description: |
        Segunda camada de screening: analisa o contexto semântico da descrição da transação
        usando LLM (via API coaf-analyzer) para determinar se o keyword match é realmente suspeito
        ou um falso positivo.

        Recebe a transação que já teve keyword match e avalia se o contexto justifica
        comunicação ao COAF.

        **Classificações possíveis:**
        - SUSPICIOUS — Contexto confirma suspeita, requer comunicação
        - NOT_SUSPICIOUS — Falso positivo, contexto não justifica alerta
        - INCONCLUSIVE — LLM não conseguiu determinar, requer revisão do analista
      operationId: evaluateContextualScreening
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContextualScreeningRequest'
            examples:
              suspicious:
                summary: "Transação com keyword em contexto suspeito"
                value:
                  transactionId: "TX-001"
                  ruleId: "CONTEXTUAL_SCREENING"
                  description: "transferencia para lavagem de dinheiro internacional"
                  matchedKeyword: "lavagem"
              falsePositive:
                summary: "Keyword em contexto inofensivo"
                value:
                  transactionId: "TX-002"
                  ruleId: "CONTEXTUAL_SCREENING"
                  description: "pagamento lavagem carro no lava-jato"
                  matchedKeyword: "lavagem"
      responses:
        '200':
          description: Resultado da análise contextual
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContextualScreeningResponse'
              examples:
                suspicious:
                  summary: Contexto confirma suspeita
                  value:
                    classification: "SUSPICIOUS"
                    confidence: 0.92
                    reason: "Descrição menciona transferência internacional associada a lavagem de dinheiro"
                    requiresAnalystReview: false
                falsePositive:
                  summary: Falso positivo identificado
                  value:
                    classification: "NOT_SUSPICIOUS"
                    confidence: 0.88
                    reason: "Contexto indica serviço de lava-jato automotivo"
                    requiresAnalystReview: false
                inconclusive:
                  summary: Requer revisão do analista
                  value:
                    classification: "INCONCLUSIVE"
                    confidence: 0.55
                    reason: "Contexto ambíguo, não foi possível determinar com confiança"
                    requiresAnalystReview: true
        '400':
          description: "Campos obrigatórios ausentes"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/rules/contextual-screening/decisions:
    post:
      tags: [Contextual Screening]
      summary: Registrar decisão do analista
      description: |
        Permite ao analista registrar sua decisão final sobre uma transação
        que passou pelo screening contextual. Usado quando a classificação
        é INCONCLUSIVE ou quando o analista discorda da classificação da LLM.

        **Decisões possíveis:**
        - APPROVE — Analista confirma que a transação é suspeita (comunicar COAF)
        - REJECT — Analista determina que é falso positivo (não comunicar)
      operationId: registerAnalystDecision
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalystDecisionRequest'
            examples:
              approve:
                summary: "Analista aprova (confirma suspeita)"
                value:
                  transactionId: "TX-001"
                  ruleId: "CONTEXTUAL_SCREENING"
                  analystDecision: "APPROVE"
              reject:
                summary: "Analista rejeita (falso positivo)"
                value:
                  transactionId: "TX-002"
                  ruleId: "CONTEXTUAL_SCREENING"
                  analystDecision: "REJECT"
      responses:
        '200':
          description: Decisão registrada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalystDecisionResponse'
              examples:
                registered:
                  value:
                    transactionId: "TX-001"
                    ruleId: "CONTEXTUAL_SCREENING"
                    analystDecision: "APPROVE"
                    registeredAt: "2025-01-15T10:45:00Z"
        '400':
          description: "Campos obrigatórios ausentes"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    Decision:
      type: string
      description: |
        Resultado da avaliação do Decision Engine.
        - **ALERT**: Todas as condições satisfeitas → ação executada
        - **IGNORE**: Pelo menos uma condição falhou → sem ação
        - **REVIEW**: (futuro) Enviar para revisão manual
        - **BLOCK**: (futuro) Bloquear transação
      enum: [ALERT, IGNORE, REVIEW, BLOCK]
      example: ALERT
    AlertStatus:
      type: string
      description: |
        Status do alerta com state machine.
        OPEN → UNDER_REVIEW → CLOSED | FALSE_POSITIVE (terminais).
      enum: [OPEN, UNDER_REVIEW, CLOSED, FALSE_POSITIVE]
      example: OPEN
...
    Action:
      type: string
      description: |
        Ação a ser executada quando a decisão é ALERT.
        - **GENERATE_ALERT**: Cria alerta no Alert Context
        - **IGNORE**: Nenhuma ação (apenas registro)
        - **REVIEW**: (futuro) Enviar para revisão
        - **BLOCK**: (futuro) Bloquear
      enum: [GENERATE_ALERT, IGNORE, REVIEW, BLOCK]
      example: GENERATE_ALERT
...
    UpdateAlertStatusRequest:
      type: object
      description: Request para transicionar o status de um alerta.
      required: [status]
      properties:
        status:
          $ref: '#/components/schemas/AlertStatus'
          description: |
            Novo status desejado. Deve respeitar a state machine:
            OPEN → UNDER_REVIEW → CLOSED | FALSE_POSITIVE

    # ═══════════════════════════════════════════════════════════════════════════════
    # RESPONSE DTOs
    # ═══════════════════════════════════════════════════════════════════════════════
...
    DecisionExecutionResponse:
      type: object
      description: |
        Registro imutável de uma execução do Decision Engine.
        Inclui todos os dados necessários para auditoria e compliance (COAF/BACEN).
        Nunca alterado ou deletado após criação.
      properties:
        id:
          type: string
          format: uuid
          description: Identificador único da execução
        transactionId:
          type: string
          description: ID da transação PIX que originou a avaliação
          example: "TX-001"
        ruleId:
          type: string
          format: uuid
          description: ID da regra que foi avaliada
        decision:
          $ref: '#/components/schemas/Decision'
          description: Decisão tomada (ALERT ou IGNORE no MVP)
        actions:
          type: array
          description: Ações executadas como resultado (vazio se IGNORE)
          items:
            $ref: '#/components/schemas/Action'
        facts:
          type: object
          description: |
            Fatos avaliados nesta execução (contexto completo da decisão).
            Mapa de factName → valor tipado.
          additionalProperties: true
          example:
            keywordMatched: true
            customerRisk: "AR"
        matchedExpressions:
          type: array
          description: Expressões que foram satisfeitas
          items:
            $ref: '#/components/schemas/ExpressionEvaluationResponse'
        failedExpressions:
          type: array
          description: Expressões que falharam
          items:
            $ref: '#/components/schemas/ExpressionEvaluationResponse'
        configurationVersion:
          type: integer
          description: Versão da configuração que foi usada na avaliação
          example: 3
        executionTimeMs:
          type: integer
          format: int64
          description: Tempo total de execução em milissegundos
          example: 12
        traceId:
          type: string
          description: |
            Identificador de rastreamento para correlação com logs.
            Permite buscar esta execução pelo traceId em sistemas de observabilidade.
          example: "trace-uuid-123"
        timestamp:
          type: string
          format: date-time
          description: Momento em que a decisão foi tomada (ISO-8601)
          example: "2025-01-15T10:35:00Z"
    PagedDecisionExecutionResponse:
      type: object
      description: Resultado paginado de execuções de decisão.
      properties:
        content:
          type: array
          description: Execuções na página atual
          items:
            $ref: '#/components/schemas/DecisionExecutionResponse'
        page:
          type: integer
          description: Número da página atual (0-indexed)
          example: 0
        size:
          type: integer
          description: Tamanho da página
          example: 20
        totalElements:
          type: integer
          format: int64
          description: Total de execuções que satisfazem o filtro
          example: 42
        totalPages:
          type: integer
          description: Total de páginas disponíveis
          example: 3
    AlertResponse:
      type: object
      description: |
        Alerta gerado automaticamente pelo Decision Engine.
        Criado quando decision=ALERT e action=GENERATE_ALERT.
        Idempotente: apenas 1 alerta por (transactionId, ruleId).
      properties:
        id:
          type: string
          format: uuid
          description: Identificador único do alerta
        transactionId:
          type: string
          description: ID da transação que originou o alerta
          example: "TX-001"
        ruleId:
          type: string
          format: uuid
          description: ID da regra que gerou o alerta
        customerId:
          type: string
          description: ID do cliente associado à transação
          example: "CUST-42"
        facts:
          type: object
          description: Fatos avaliados que levaram ao alerta
          additionalProperties: true
          nullable: true
          example:
            keywordMatched: true
            customerRisk: "AR"
        configurationVersion:
          type: integer
          description: Versão da configuração que estava ativa
          nullable: true
          example: 3
        traceId:
          type: string
          description: TraceId para correlação com a DecisionExecution
          nullable: true
          example: "trace-uuid-123"
        actions:
          type: array
          description: Ações que foram executadas
          items:
            type: string
          nullable: true
          example: ["GENERATE_ALERT"]
        explanation:
          type: object
          description: Dados resumidos da explicação da decisão
          additionalProperties: true
          nullable: true
        status:
          $ref: '#/components/schemas/AlertStatus'
          description: Status atual do alerta (segue state machine)
        createdAt:
          type: string
          format: date-time
          description: Quando o alerta foi criado
          example: "2025-01-15T10:35:00Z"
        updatedAt:
          type: string
          format: date-time
          description: Última atualização de status
          example: "2025-01-15T10:35:00Z"

    # ═══════════════════════════════════════════════════════════════════════════════
    # SCREENING CONTEXT — Keyword & Contextual Screening
    # ═══════════════════════════════════════════════════════════════════════════════
...
    ContextualScreeningRequest:
      type: object
      description: |
        Request para análise contextual de uma transação que já teve keyword match.
        Envia a descrição + keyword para avaliação via LLM (coaf-analyzer).
      required: [transactionId, description, matchedKeyword]
      properties:
        transactionId:
          type: string
          description: "ID da transação que teve keyword match"
          example: "TX-001"
        ruleId:
          type: string
          description: "Código da regra (padrão CONTEXTUAL_SCREENING se omitido)"
          example: "CONTEXTUAL_SCREENING"
        description:
          type: string
          description: "Descrição completa da transação PIX"
          example: "pagamento lavagem carro no lava-jato"
        matchedKeyword:
          type: string
          description: "Termo que foi detectado pelo Keyword Screening"
          example: "lavagem"
    ContextualScreeningResponse:
      type: object
      description: |
        Resultado da análise contextual via LLM.
        Indica se o contexto confirma a suspeita ou é falso positivo.
      properties:
        classification:
          type: string
          description: |
            Classificação da LLM:
            - SUSPICIOUS — Contexto confirma suspeita
            - NOT_SUSPICIOUS — Falso positivo
            - INCONCLUSIVE — Requer revisão humana
          enum: [SUSPICIOUS, NOT_SUSPICIOUS, INCONCLUSIVE]
          example: "NOT_SUSPICIOUS"
        confidence:
          type: number
          format: double
          description: "Grau de confiança da classificação (0.0 a 1.0)"
          minimum: 0.0
          maximum: 1.0
          example: 0.88
        reason:
          type: string
          description: "Justificativa em linguagem natural da classificação"
          example: "Contexto indica serviço de lava-jato automotivo"
        requiresAnalystReview:
          type: boolean
          description: "Se a transação precisa de revisão manual pelo analista"
          example: false
    AnalystDecisionRequest:
      type: object
      description: Request para registrar a decisão final do analista sobre uma transação.
      required: [transactionId, analystDecision]
      properties:
        transactionId:
          type: string
          description: "ID da transação sendo decidida"
          example: "TX-001"
        ruleId:
          type: string
          description: "Código da regra (padrão CONTEXTUAL_SCREENING se omitido)"
          example: "CONTEXTUAL_SCREENING"
        analystDecision:
          type: string
          description: |
            Decisão do analista:
            - APPROVE — Confirma suspeita (comunicar COAF)
            - REJECT — Falso positivo (não comunicar)
          enum: [APPROVE, REJECT]
          example: "APPROVE"
    AnalystDecisionResponse:
      type: object
      description: Confirmação de que a decisão do analista foi registrada.
      properties:
        transactionId:
          type: string
          description: "ID da transação"
          example: "TX-001"
        ruleId:
          type: string
          description: "Código da regra"
          example: "CONTEXTUAL_SCREENING"
        analystDecision:
          type: string
          description: "Decisão registrada (APPROVE ou REJECT)"
          example: "APPROVE"
        registeredAt:
          type: string
          format: date-time
          description: "Momento em que a decisão foi registrada"
          example: "2025-01-15T10:45:00Z"

    # ═══════════════════════════════════════════════════════════════════════════════
    # ERROR RESPONSE
    # ═══════════════════════════════════════════════════════════════════════════════
    ErrorResponse:
      type: object
      description: |
        Formato padronizado de erro para todos os endpoints.
        Inclui timestamp, status HTTP, código de erro e mensagem descritiva.
      properties:
        timestamp:
          type: string
          format: date-time
          description: Momento do erro (ISO-8601)
          example: "2025-01-15T10:30:00Z"
        status:
          type: integer
          description: Código HTTP do erro
          example: 422
        error:
          type: string
          description: Descrição curta do tipo de erro HTTP
          example: "Unprocessable Entity"
        message:
          type: string
          description: Mensagem detalhada do erro (legível)
          example: "Fato 'unknownFact' não encontrado no catálogo"
        details:
          type: string
          description: "Detalhes adicionais (ex.: campos com erro de validação)"
          nullable: true
          example: "expressions[0].factName: não pode ser vazio; createdBy: é obrigatório"